robot_script/requirements.txt
deep_sort_realtime==1.3.2
firebase_admin==7.1.0
numpy==2.3.2
opencv_python==4.12.0.88
Pillow==11.3.0
pynmea2==1.19.0
pyrealsense2==2.56.5.9235
pyserial==3.5
python-dotenv==1.1.1
python-socketio==5.13.0
Requests==2.32.5
torch==2.5.1+cu121
transformers==4.54.0
ultralytics==8.3.168
uvicorn==0.35.0
